api-connectors/requirements.txt
# API Connectors Dependencies

# Database
psycopg2-binary==2.9.10

# HTTP Clients
requests==2.31.0
httpx==0.25.2

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# HubSpot
hubspot-api-client==8.2.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0

# Data processing
pandas==2.1.3
numpy==1.26.4

# Async
aiohttp==3.9.1
asyncio-throttle==1.0.2
